Disallow assigning value-type arrays to json. Fix #4348
