+++ modules/ballerina-test/src/test/resources/logging.properties
@@ -34,4 +34,4 @@ ballerina.useParentHandlers=false
 
 # JUL root logger
 .handlers=org.ballerinalang.test.utils.TestLogHandler
-.level=INFO
+.level=OFF
