--- modules/ballerina-test/src/test/resources/test-src/types/jsontype/json-literal-negative.bal
@@ -1,11 +1,0 @@
-function testJsonArrayWithUnsupportedtypes() (json) {
-    datatable dt;
-    json j = ["a", "b", "c", dt];
-    return j;
-}
-
-function testJsonInitWithUnsupportedtypes() (json) {
-    datatable dt;
-    json j = {"name":"Supun", "value":dt};
-    return j;
-}
