+++ modules/ballerina-test/src/test/resources/test-src/types/jsontype/json-value-negative.bal
@@ -1,12 +1,35 @@
-function testGetFromNull () (string) {
-    json j2 = {age:43, name:null};
-    string value;
-    value, _ = (string)j2.name.fname;
-    return value;
+function testJsonArrayWithUnsupportedtypes() (json) {
+    datatable dt;
+    json j = ["a", "b", "c", dt];
+    return j;
 }
 
-function testAddToNull () (json) {
-    json j = {name:"Supun", address:null};
-    j.address.country = "SriLanka";
+function testJsonInitWithUnsupportedtypes() (json) {
+    datatable dt;
+    json j = {"name":"Supun", "value":dt};
+    return j;
+}
+
+function testIntArrayToJsonAssignment() (json) {
+    int[] a = [1, 5, 9];
+    json j = a;
+    return j;
+}
+
+function testFloatArrayToJsonAssignment() (json) {
+    float[] f = [1.3, 5.4, 9.4];
+    json j = f;
+    return j;
+}
+
+function testStringArrayToJsonAssignment() (json) {
+    string[] s = ["apple", "orange"];
+    json j = s;
+    return j;
+}
+
+function testBooleanArrayToJsonAssignment() (json) {
+    boolean[] b = [true, true, false];
+    json j = b;
     return j;
 }
