+++ modules/ballerina-test/src/test/resources/test-src/types/jsontype/json-value.bal
@@ -330,3 +330,16 @@ function testJsonToJsonArrayInvalidCasting () (json[][][], TypeCastError) {
 
     return j2, e;
 }
+
+function testGetFromNull () (string) {
+    json j2 = {age:43, name:null};
+    string value;
+    value, _ = (string)j2.name.fname;
+    return value;
+}
+
+function testAddToNull () (json) {
+    json j = {name:"Supun", address:null};
+    j.address.country = "SriLanka";
+    return j;
+}
